chore: refactoring of itest code
